+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.1.1 - 2017-08-24
+### Added
+- Added a submit button field.
+- Added ability to specify wrap HTML tag for a field.
+### Changed
+- Enhanced `selected` and `checked` methods.
+
+### Changed
+- Slugify field `type` with underscores intead of dashes.
+
 ## 0.1.0 - 2017-08-13
 ### Added
 - `Field` class